requirements.txt
streamlit==1.28.1
openai==1.95.1
python-dotenv==1.0.0
thefuzz==0.19.0
pytest==7.4.0
sentence-transformers==2.2.2
numpy==1.24.3 